apiservice/requirements.txt
azure-storage-blob==1.5.0
azure-storage-queue
azure-cosmosdb-table
